requirements.txt
flask==3.0.3
flask-cors==5.0.0
openai-whisper==20240930
requests==2.32.3
python-dotenv==1.0.1
